Correções I

+++ src/main/java/org/ufla/spark/rec_inf_tp2/funcoes/FunParaTuple2.java
@@ -2,6 +2,7 @@ package org.ufla.spark.rec_inf_tp2.funcoes;
 
 import org.apache.spark.api.java.function.MapFunction;
 import org.apache.spark.sql.Row;
+
 import scala.Tuple2;
 
 /**
@@ -36,7 +37,7 @@ public class FunParaTuple2 implements MapFunction<Row, Tuple2<Object, Object>> {
 	@Override
 	public Tuple2<Object, Object> call(Row row) throws Exception {
 		Object predicao = row.get(colunaPredicao);
-		Object label = row.get(colunaLabel);
+		Object label = (Double) ((Integer) row.get(colunaLabel)).doubleValue();
 		return new Tuple2<Object, Object>(predicao, label);
 	}
